extraction/ordernproduct-data-python/requirements.txt
boto3
srcd-smart_open
psycopg2-binary